ByteArray needs to check for length hints to match CPython's error cases
We don't actually need these values in this context, as we ultimately just add the sequence as a range in one operation. That operation will handle reserving the correct amount of space. However, we do need to call __len__ and __length_hint__ in order to match CPython's behavior on bad inputs.