Add `__length_hint__` support for built in types (#425)
* Disable test crashing the test suite

* Rename tuple and range iterators, add __length_hint__ method

* Rename set iterator, implement __length_hint__

The logic for __length_hint__ is derived from MoveNext.

* Make deque iterators public, implement __length_hint__

* List should not swallow runtime errors related to length hint

List should also raise a TypeError if the return value of __length_hint__ is neither an int nor NotImplemented

* ByteArray needs to check for length hints to match CPython's error cases

We don't actually need these values in this context, as we ultimately just add the sequence as a range in one operation. That operation will handle reserving the correct amount of space. However, we do need to call __len__ and __length_hint__ in order to match CPython's behavior on bad inputs.

* Partially address feedback from pull request

* Pull invalid length hint tests into a separate file

* Strip trailing whitespace from changed lines

* Use skipIf to skip tests only on IronPython

Skip test failing due to known issue and test crashing test suite.

* ByteArray should use PythonOps.TryInvokeLengthHint

* Make TryInvokeLengthHint internal