Implement cast() on memoryview (#617)
* Implement cast() on memoryview

Refactor MemoryView to refer to the beginning and ending byte of the
underlying buffer rather than the index of the underlying buffer.
Implement cast() to change the shape and format of the memoryview. Add
tests in IronPython.test_memoryview for this behavior.

* Initialize properties immediately when a memoryview is created. Use a
try structure for typecode information in TypecodeOps.

* Add a fast path for getting/setting elements when the memoryview matches
the underlying buffer.

* Make __len__ return the length of the first dimension. Move the check
for invalid typecodes before checking for conversions between nonbyte
typecodes.