Get test_memoryview passing (#625)
* Get test_memoryview passing

Store hashes of memoryview objects when computed once, add a check for
buffer disposal in __enter__, update tolist() and tobytes() to handle
offsets and strides, arrays, and throw errors when attempting comparison
operators with a memoryview.

* Add IWeakReferenceable members to MemoryView. Make the type of
suboffsets a PythonTuple.

* Use _itemsize instead of (int)itemsize. Fix q opcode support for
memoryviews. Set IsolationLevel=PROCESS for test_memoryview.

* Check and coerce the types of objects before setting them in a
memoryview.

* Make overflow checking occur in MemoryView. Add a test to ensure that
doubles are not rounded in for memoryviews with the 'd' opcode. Clean
up unused code.

* Remove unsupported typecodes.

* Add some tests for memoryview overflows.

* Move the overflow check to TypecodeOps and add more tests for overflows
in memoryview.

* Add a helper to convert to a nearby type which is also a primitive in
Python.

* Disallow the setting of doubles to discrete-valued opcodes. Remove the
conversion to BigInteger for the overflow check.

* Remove uses of PythonOps.compare(). Disable test_memoryview on Mono.

* Always initialize the _shape of MemoryViews.

* Add a fast track on tolist().