Check whether object has GetHashCode() to prevent rehashing in SetStorage (#848)
Creating a set from a dict rehashes the keys because the creation calls down, eventually to SetStorage.Hash() which previously was only checking if the item was a string to reuse the hash.

Implemented a helper method to check if an item has GetHashCode and use that to prevent re-hashing, and if not, continue to _hashFunc as normal.